reorganise LMV_ExtendedTests
- moved headless login function to a generic file.
