+++ Assets/Tests/LMV_ExtendedTests/LMV_ImageFetchingTests.cs
@@ -12,7 +12,7 @@ using OpenMetaverse.Http;
 using Plugins.ObjectPool;
 using Tests;
 
-namespace Raindrop.Tests
+namespace Raindrop.Tests.LMV_ExtendedTests
 {
     [TestFixture]
     public class LMVLibraryImageFetchingTests
@@ -36,51 +36,6 @@ namespace Raindrop.Tests
             instance = new RaindropInstance(new GridClient());
         }
         
-        // Do login using the LMV library only.
-        void LoginHeadless(int userIdx = 0, string startLocation = "Hooper")
-        {
-            /* hack: I added this,
-             * to prevent the library from throwing the runtime exception :
-             *
-             * Unhandled log message: '[Error] 22:34:22 [ERROR] - <TanukiDEV
-             * Resident>: Setting server side baking failed'. Use
-             * UnityEngine.TestTools.LogAssert.Expect   */
-             
-            instance.Client.Settings.SEND_AGENT_APPEARANCE = false;
-
-            var fullUsername = Secrets.GridUsers[userIdx];
-            var password = Secrets.GridPass[userIdx];
-            Assert.IsFalse(string.IsNullOrWhiteSpace(fullUsername),
-                "LMVTestAgentUsername is empty. " +
-                "Live NetworkTests cannot be performed.");
-            Assert.IsFalse(string.IsNullOrWhiteSpace(password),
-                "LMVTestAgentPassword is empty. " +
-                "Live NetworkTests cannot be performed.");
-            var username = fullUsername.Split(' ');
-
-            // Connect to the grid
-            string startLoc = 
-                NetworkManager.StartLocation(startLocation, 179, 18, 32);
-            Debug.Log($"Logging in " +
-                      $"User: {fullUsername}, " +
-                      $"Loc: {startLoc}");
-            bool loginSuccessful = instance.Client.Network.Login(
-                username[0],
-                username[1],
-                password,
-                "Unit Test Framework",
-                startLoc,
-                "raindropcafeofficial@gmail.com");
-            Assert.IsTrue(loginSuccessful, 
-                $"Client failed to login, reason: " +
-                $"{instance.Client.Network.LoginMessage}");
-            Debug.Log("Grid returned the login message: " 
-                      + instance.Client.Network.LoginMessage);
-
-            Assert.IsTrue(
-                instance.Client.Network.Connected, 
-                "Client is not connected to the grid");
-        }
 
         /* DownloadManager variant of 'wtf' WebRequest.
          * (Used to) Fail in unity editor, android target, net standard 2.0
@@ -353,7 +308,7 @@ namespace Raindrop.Tests
         [Test]
         public IEnumerator LoginAndDownloadJ2P()
         {
-            LoginHeadless();
+            Helpers.LoginHeadless(instance, 0, "Hooper");
 
             yield return new WaitForSeconds(15);
